backend/pyproject.toml
[tool.poetry]
name = "video-agent-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"
pydantic = "^2.5.3"
langchain = "^0.1.1"
langchain-community = "^0.0.13"
langchain-core = "^0.1.12"
langchain-openai = "^0.0.2"
langgraph = "^0.0.20"
faiss-cpu = "^1.7.4"
chromadb-client = "^0.4.22"
sentence-transformers = "^2.2.2"
speechrecognition = "^3.10.1"
pydub = "^0.25.1"
openai-whisper = "^20231117"
ffmpeg-python = "^0.2.0"
torch = "^2.1.2"
torchvision = "^0.16.2"
torchaudio = "^2.1.2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
